Create RepositoryFacadeBuilder
