+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/ChatDomainImpl.kt
@@ -44,6 +44,7 @@ import io.getstream.chat.android.livedata.extensions.users
 import io.getstream.chat.android.livedata.model.ChannelConfig
 import io.getstream.chat.android.livedata.model.SyncState
 import io.getstream.chat.android.livedata.repository.RepositoryFacade
+import io.getstream.chat.android.livedata.repository.RepositoryFacadeBuilder
 import io.getstream.chat.android.livedata.repository.RepositoryFactory
 import io.getstream.chat.android.livedata.request.AnyChannelPaginationRequest
 import io.getstream.chat.android.livedata.request.QueryChannelPaginationRequest
@@ -68,7 +69,6 @@ import java.util.Date
 import java.util.InputMismatchException
 import java.util.UUID
 import java.util.concurrent.ConcurrentHashMap
-import kotlin.collections.set
 
 private val CHANNEL_CID_REGEX = Regex("^!?[\\w-]+:!?[\\w-]+$")
 private const val MESSAGE_LIMIT = 30
@@ -246,7 +246,12 @@ internal class ChatDomainImpl internal constructor(
 
         database = db ?: createDatabase(appContext, user, offlineEnabled)
 
-        repos = RepositoryFacade.create(factory = RepositoryFactory(database, user), scope = scope, defaultConfig = defaultConfig)
+        repos = RepositoryFacadeBuilder {
+            factory = RepositoryFactory(database, user)
+            coroutineScope = scope
+            defaultConfig = this@ChatDomainImpl.defaultConfig
+        }
+            .build()
 
         // load channel configs from Room into memory
         initJob = scope.async {
